# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Polymer Bridge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BridgeContractsBase = {
  chainId: number;
  contractAddress: string;
  rpcUrl: string;
  explorerUrl: string;
};

export type BridgeContractsPlume = {
  chainId: number;
  contractAddress: string;
  rpcUrl: string;
  explorerUrl: string;
};

export interface BridgeContracts {
  base: BridgeContractsBase;
  plume: BridgeContractsPlume;
  /** CrossL2Prover contract address (same on all chains) */
  polymerProver: string;
}

export type BridgeTransactionStatus =
  (typeof BridgeTransactionStatus)[keyof typeof BridgeTransactionStatus];

export const BridgeTransactionStatus = {
  pending: "pending",
  proving: "proving",
  proven: "proven",
  claimed: "claimed",
  failed: "failed",
} as const;

export interface BridgeTransaction {
  id: number;
  txHash: string;
  srcChainId: number;
  destChainId: number;
  sender: string;
  recipient: string;
  /** Amount in wei as string */
  amount: string;
  nonce: string;
  status: BridgeTransactionStatus;
  blockNumber?: number | null;
  logIndex?: number | null;
  globalLogIndex?: number | null;
  polymerJobId?: string | null;
  proof?: string | null;
  claimTxHash?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBridgeTransactionBody {
  txHash: string;
  srcChainId: number;
  destChainId: number;
  sender: string;
  recipient: string;
  amount: string;
  nonce: string;
  blockNumber?: number;
  logIndex?: number;
  globalLogIndex?: number;
}

export interface RequestProofBody {
  /** Transaction hash of the deposit on the source chain */
  txHash: string;
  /** Source chain ID */
  srcChainId: number;
  /** Block number of the deposit transaction */
  blockNumber: number;
  /** Global log index of the Deposited event in the block */
  globalLogIndex: number;
}

export type ProofJobResponseStatus =
  (typeof ProofJobResponseStatus)[keyof typeof ProofJobResponseStatus];

export const ProofJobResponseStatus = {
  pending: "pending",
  complete: "complete",
  error: "error",
} as const;

export interface ProofJobResponse {
  jobId: string;
  txHash: string;
  status: ProofJobResponseStatus;
}

export type ProofStatusResponseStatus =
  (typeof ProofStatusResponseStatus)[keyof typeof ProofStatusResponseStatus];

export const ProofStatusResponseStatus = {
  pending: "pending",
  complete: "complete",
  error: "error",
} as const;

export interface ProofStatusResponse {
  jobId: string;
  status: ProofStatusResponseStatus;
  /** Hex-encoded proof bytes ready for on-chain submission */
  proof?: string | null;
}

export interface BridgeStats {
  totalTransactions: number;
  /** Total bridged amount in wei as string */
  totalVolume: string;
  pendingCount: number;
  completedCount: number;
  baseToPlume: number;
  plumeToBase: number;
}

export interface ApiError {
  error: string;
  message: string;
}

export type GetRecentBridgeTransactionsParams = {
  limit?: number;
};